requirements.txt
opencv-python>=4.5.0
pytesseract>=0.3.10
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
fhirpy>=1.4.0
Flask>=2.3.0
Flask-RESTful>=0.3.9
python-dotenv>=1.0.0
cryptography>=41.0.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
spacy>=3.5.0
nltk>=3.8
pydicom>=2.4.0
pillow>=10.0.0
requests>=2.31.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
python-logging-loki>=0.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
gunicorn>=21.0.0
click>=8.1.0
pyyaml>=6.0
# Tesseract-OCR required: Windows - https://github.com/UB-Mannheim/tesseract/wiki, Linux: apt install tesseract-ocr
# TensorFlow optional - install separately if needed: pip install tensorflow-cpu
# For GPU support: pip install tensorflow-gpu
